+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
-import { ProductsService } from '../../services/products.service';
 import { ProductModel } from '../../models/product.model';
+
+import { ProductsService } from '../../services/products.service';
+import { CartService } from '../../../cart/services/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -12,10 +14,14 @@ export class ProductListComponent implements OnInit {
 
   products: Array<ProductModel>;
 
-  constructor(private productsService: ProductsService) { }
+  constructor(private productsService: ProductsService, private cartService: CartService) { }
 
   ngOnInit() {
     this.products = this.productsService.getProducts();
+  }
+
+  onBuy(product: ProductModel): void {
+    this.cartService.publishData(product);
   }
 
 }